Bash Script to calculate the strength of schedule
The project was to extract the relevant data from an HTML file.
Implementation consisted of SED for extraction, along with GREP and
other supporting techniques. The calculation was based on number of
wins, losses and ties. The total number of wins with ties factored in
and divided by theoretically possible wins results the team strength.
(BASH script).